+++ src/Pages/index.js
@@ -1,0 +1,34 @@
+import { useTodoApp } from "../components/index.js";
+import { useFilter } from "../hooks/index.js";
+
+export const Pages = () => {
+  const {
+    TodoInput,
+    TodoList,
+    tasks,
+    addTask,
+    deleteTask,
+    toggleComplete,
+    clearTasks,
+    setFilter,
+    setSearchQuery,
+  } = useTodoApp();
+
+  const FilterComponent = useFilter({ setFilter, setSearchQuery });
+
+  return (
+    <div>
+      <h1>To-Do List</h1>
+      <TodoInput addTask={addTask} />
+      {FilterComponent}
+      <TodoList
+        tasks={tasks}
+        deleteTask={deleteTask}
+        toggleComplete={toggleComplete}
+      />
+      <button className="clear-tasks" onClick={clearTasks}>
+        Clear All Tasks
+      </button>
+    </div>
+  );
+};
