+++ src/components/TaskInput/index.js
@@ -1,10 +1,11 @@
 import React, { useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 function TodoInput({ addTask }) {
   const [description, setDescription] = useState("");
   const [priority, setPriority] = useState("Low");
-
+  const notify = () => toast.error("Please add a task description.");
   const handleSubmit = (e) => {
     e.preventDefault();
     if (description.trim()) {
@@ -16,26 +17,42 @@ function TodoInput({ addTask }) {
       });
       setDescription("");
       setPriority("Low");
+    } else {
+      notify();
     }
   };
 
   return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-        placeholder="Add a task..."
+    <>
+      <form onSubmit={handleSubmit}>
+        <input
+          type="text"
+          value={description}
+          onChange={(e) => setDescription(e.target.value)}
+          placeholder="Add a task..."
+        />
+        <select value={priority} onChange={(e) => setPriority(e.target.value)}>
+          <option value="Low">Low</option>
+          <option value="Medium">Medium</option>
+          <option value="High">High</option>
+        </select>
+        <button className="submit" type="submit">
+          Add
+        </button>
+      </form>
+      <ToastContainer
+        position="top-center"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="colored"
       />
-      <select value={priority} onChange={(e) => setPriority(e.target.value)}>
-        <option value="Low">Low</option>
-        <option value="Medium">Medium</option>
-        <option value="High">High</option>
-      </select>
-      <button className="submit" type="submit">
-        Add
-      </button>
-    </form>
+    </>
   );
 }
