+++ src/hooks/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import Filter from "../hooks/Filter/index.js";
+
+export const useFilter = ({ setFilter, setSearchQuery }) => (
+  <Filter setFilter={setFilter} setSearchQuery={setSearchQuery} />
+);
